Refactored todo-list-section & updated home screen
